SIGH.
MSSQL doesn't do BETWEEN for dates the way anyone on earth seems to expect[1]. This is a bug that's been around since the very first version of these reports - they leave off anything on the last day of a date range.

borrowing, borrowingtotals, lending, lendingdetails were all affected.

1: http://stackoverflow.com/questions/749615/does-ms-sql-servers-between-include-the-range-boundaries

Thanks to Alex Lee for the bug report.